datakit_testbed: consolidate build_testbed_steps into sampler; add baseline.py
Two related moves:

1. ``build_testbed_steps`` + ``_sample_step_for`` move from dag.py into
   sampler.py — the function is just "build sample steps per source",
   which belongs next to the rest of the sampling code. dag.py is
   deleted (one less module, no new concept lived there).
2. ``baseline`` moves into its own module (experiments/datakit_testbed/
   baseline.py) so the control-arm wiring has a dedicated home and can
   grow new config knobs without crowding sampler.py.

Import sites and tests updated; test_dag.py renamed to
test_sampler_pipeline.py.
