+++ tests/datakit_testbed/test_sampler_pipeline.py
@@ -5,7 +5,7 @@
 
 import pytest
 
-from experiments.datakit_testbed.dag import build_testbed_steps
+from experiments.datakit_testbed.sampler import build_testbed_steps
 from marin.datakit.sources import DatakitSource, all_sources
 
 _ALL = all_sources()
